+++ statick_tool/plugins/reporting/json_reporting_plugin.py
@@ -73,25 +73,26 @@ class JsonReportingPlugin(ReportingPlugin):
 
     def write_output(self, package: Package, level: str, line: str) -> bool:
         """Write JSON output to a file."""
-        if not self.plugin_context:
-            return False
+        # By default write report to the current directory.
+        output_dir = os.getcwd()
+        if (
+            self.plugin_context
+            and "output_directory" in self.plugin_context.args
+            and self.plugin_context.args.output_directory is not None
+        ):
+            # If an output directory is specified use it for the report.
+            output_dir = os.path.join(
+                self.plugin_context.args.output_directory, package.name + "-" + level
+            )
 
-        output_dir = os.path.join(
-            self.plugin_context.args.output_directory, package.name + "-" + level
-        )
         if not os.path.exists(output_dir):
             os.mkdir(output_dir)
         if not os.path.isdir(output_dir):
             logging.error("Unable to create output directory at %s!", output_dir)
             return False
-        output_file = os.path.join(output_dir, package.name + "-" + level + ".json")
-        next_output_file = os.path.join(
+
+        output_file = os.path.join(
             output_dir, package.name + "-" + level + ".statick.json"
-        )
-        logging.warning(
-            "Output filename will change in statick v0.9.0 from %s to %s",
-            output_file,
-            next_output_file,
         )
         logging.info("Writing output to %s", output_file)
         with open(output_file, "w", encoding="utf8") as out: