+++ tests/plugins/reporting/json_reporting_plugin/test_json_reporting_plugin.py
@@ -125,4 +125,9 @@ def test_json_reporting_plugin_write_output_no_plugin_context():
         package = Package(
             "valid_package", os.path.join(os.path.dirname(__file__), "valid_package")
         )
-        assert not jrp.write_output(package, "level", "line")
+        assert jrp.write_output(package, "level", "line")
+        output_file = os.path.join(
+            os.getcwd(), package.name + "-" + "level" + ".json"
+        )
+        if os.path.exists(output_file):
+            os.remove(output_file)